feat(102): add new AEP around APIs (#354)
Adding an AEP around APIs specifically, to include more specific guidance and a
landing page for APIs.

Clarifying some guidance around the usage of sub-paths for API names: previously
this was confusing as an example existed (`api.example.com/users`) which seemed
to allow this, but the API name guidance did not allow path elements in the API 
name.

Also adding guidance for the `type` field in the `x-aep-resource` oas annotation: this was missing from
the schema annotation previously, which effectively did not require placing the 
api name anywhere in an OAS schema description.