fix(122): always use full resource path for API references (#343)
The previous guidance was ambiguous about when to use a full
resource name vs a relative one. 

This ambiguity makes it difficult for clients to read the location of 
the remote resource, as well as the ablity to resolve it as necessary.

Some forms of REST APIs already put the full resource name directly in the 
payload, so there is precedence for this approach.

Clarifying the guidance to ensure consistent behavior for APIs.