feat(0006): move GOVERNANCE.md to an AEP (#359)
GOVERNANCE.md is better as an AEP that has 
more visibility in the project (e.g. it will be generated as a page in aep.dev).

Removing an article referring to the CNCF, since
the long-term home of the project is still TBD.

Minor update to @toumorokoshi's membership to reflect
their new organization.